requirements/base.txt
Django==1.10.1
psycopg2==2.6.2
python-dateutil==2.5.3
djangorestframework==3.4.7
djangorestframework-jsonapi==2.1.1
django-cors-headers==1.2.2
django-filter==0.15.3
django-import-export
django-extra-fields
